Comment out v2 tests TestFetchMetricsSummaryHandler & TestFetchMetricsThroughputTimeseriesHandler
